LayoutTests/imported/w3c/web-platform-tests/cookie-store/cookieStore_event_delete.https.window-expected.txt


PASS cookieStore fires change event for cookie deleted by cookieStore.delete()
FAIL cookieStore does not fire change events for non-existing expired cookies assert_equals: expected 2 but got 4
